requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
streamlit-drawable-c
anvas==0.9.0
plotly==5.17.0
